Finish moving this to a more scala-like idiom.
Before conversion, there was a base class AbstractCard, and then a specific
game subclassed this in order to add a method returning the card's value in
that game.

Now, there is a Card class, and an implicit conversion within the specific
game allows the value method to be mixed in on the fly.